style(type): change folder and name
