--- src/constants/common.ts
@@ -1,34 +1,0 @@
-const ATTENTION_FOCUS = {
-  EXTRAVERTED: 'E',
-  INTROVERTED: 'I' 
-} as const;
-
-const PERCEPTION_FUNCTION = {
-  SENSING: 'S',
-  INTUITION: 'N'
-} as const;
-
-const JUDGMENT_FUNCTION = {
-  THINKING: 'T',
-  FEELING: 'F'
-} as const;
-
-const LIFESTYLE = {
-  JUDGING: 'J',
-  PERCEIVING: 'P'
-} as const;
-
-const QUESTION = {
-  ATTENTION_FOCUS: ATTENTION_FOCUS,
-  PERCEPTION_FUNCTION: PERCEPTION_FUNCTION,
-  JUDGMENT_FUNCTION: JUDGMENT_FUNCTION,
-  LIFESTYLE: LIFESTYLE
-} as const;
-
-export {
-  ATTENTION_FOCUS,
-  PERCEPTION_FUNCTION,
-  JUDGMENT_FUNCTION,
-  LIFESTYLE,
-  QUESTION,
-};
