+++ src/constants/mbti.ts
@@ -1,4 +1,29 @@
-import { ATTENTION_FOCUS, JUDGMENT_FUNCTION, LIFESTYLE, PERCEPTION_FUNCTION, QUESTION } from "src/constants/common";
+const ATTENTION_FOCUS = {
+  EXTRAVERTED: 'E',
+  INTROVERTED: 'I' 
+} as const;
+
+const PERCEPTION_FUNCTION = {
+  SENSING: 'S',
+  INTUITION: 'N'
+} as const;
+
+const JUDGMENT_FUNCTION = {
+  THINKING: 'T',
+  FEELING: 'F'
+} as const;
+
+const LIFESTYLE = {
+  JUDGING: 'J',
+  PERCEIVING: 'P'
+} as const;
+
+const RESULT = {
+  ATTENTION_FOCUS: ATTENTION_FOCUS,
+  PERCEPTION_FUNCTION: PERCEPTION_FUNCTION,
+  JUDGMENT_FUNCTION: JUDGMENT_FUNCTION,
+  LIFESTYLE: LIFESTYLE
+} as const;
 
 type ATTENTION_FOCUS_TYPE = typeof ATTENTION_FOCUS[keyof typeof ATTENTION_FOCUS];
 type PERCEPTION_FUNCTION_TYPE = typeof PERCEPTION_FUNCTION[keyof typeof PERCEPTION_FUNCTION];
@@ -7,7 +32,7 @@ type LIFESTYLE_TYPE = typeof LIFESTYLE[keyof typeof LIFESTYLE];
 
 type MBTI_SINGLE_TEMPLATE_TYPE = ATTENTION_FOCUS_TYPE | PERCEPTION_FUNCTION_TYPE | JUDGMENT_FUNCTION_TYPE | LIFESTYLE_TYPE;
 
-type QUESTION_TYPE = keyof typeof QUESTION;
+type QUESTION_TYPE = keyof typeof RESULT;
 type ANSWER_TYPE<T extends MBTI_SINGLE_TEMPLATE_TYPE> = T;
 type QUESTION_SENDER_TYPE = {
   type: QUESTION_TYPE,
@@ -38,6 +63,11 @@ type QUESTION_AND_ANSWER_TYPE<T extends MBTI_SINGLE_TEMPLATE_TYPE> = {
 type MBTI_TYPE = `${ATTENTION_FOCUS_TYPE}${PERCEPTION_FUNCTION_TYPE}${JUDGMENT_FUNCTION_TYPE}${LIFESTYLE_TYPE}`;
 
 export {
+  ATTENTION_FOCUS,
+  JUDGMENT_FUNCTION,
+  LIFESTYLE,
+  PERCEPTION_FUNCTION,
+  RESULT as QUESTION,
   ATTENTION_FOCUS_TYPE,
   PERCEPTION_FUNCTION_TYPE,
   JUDGMENT_FUNCTION_TYPE,
