--- src/constants/test.ts
@@ -1,5 +1,0 @@
-export interface IBaseTestService{
-  addQuestionAndAnswers(question: any, answers: any[]): void;
-  generateRandomTest(numTests: number): any[];
-  calculateResult(choices: any[]): any;
-}
