+++ src/interfaces/test.interface.ts
@@ -1,0 +1,5 @@
+export interface IBaseTestService{
+  addQuestionAndAnswers(question: any, answers: any): void;
+  generateRandomTest(numTests: number): any[];
+  calculateResult(answers: any): any;
+}
